requirements.txt
Flask==2.0.1
flask-cors==3.0.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
pandas==1.3.3
openpyxl==3.0.9
requests==2.26.0
python-barcode==0.13.1
qrcode==7.3.1
pillow==8.3.2
pythonping==1.1.0
dnspython==2.1.0
pyyaml==6.0
google-generativeai==0.3.2
python-dotenv==0.19.1
pytesseract==0.3.8
python-docx==0.8.11
fpdf==1.7.2
opencv-python==4.5.3.56
Verbalexpressions==0.0.2
